requirements.txt
PyGithub==2.1.1
requests==2.31.0
google-generativeai==0.3.2
python-dotenv==1.0.0
python-telegram-bot==20.7
apscheduler==3.10.4